[Fix][WRS-1860] Prevent publishing of connector without explicitly specified allowedDomains options
