+++ .gitignore
@@ -1,7 +1,6 @@
 node_modules
 out
 yarn-error.log
-publish
 .npmrc
 
 #Mac OS
