+++ src/connector-cli/src/commands/publish/steps/get-request-url.ts
@@ -1,0 +1,8 @@
+export function getRequestUrl(baseUrl: string, environment: string) {
+  const connectorEndpointBaseUrl = new URL(baseUrl);
+  if (!connectorEndpointBaseUrl.pathname.endsWith('/')) {
+    connectorEndpointBaseUrl.pathname += '/';
+  }
+  connectorEndpointBaseUrl.pathname += `api/experimental/environment/${environment}/connectors`;
+  return connectorEndpointBaseUrl.toString();
+}
